refactor: LottoCount -> LottoCountDto로 수정
