+++ README.md
@@ -14,7 +14,6 @@
 
 ## 구현할 기능 목록
 * [x] 구입 금액을 입력하는 기능
-    * [x] **예외** null값이나 빈 값이 들어오는 경우
     * [x] **예외** 구입 금액이 정수가 아닌 경우
     * [x] **예외** 천 원 단위가 아닌 경우
     * [x] **예외** 한 장도 살수 없는 금액인 경우
