+++ src/main/java/lotto/controller/LottoController.java
@@ -2,15 +2,13 @@ package lotto.controller;
 
 import java.io.IOException;
 
-import lotto.domain.LottoCount;
 import lotto.utils.InputUtil;
 import lotto.view.InputView;
 import lotto.view.OutputView;
 
 public class LottoController {
 	public static void run() {
-		LottoCount lc = new LottoCount(readMoney());
-		// LottoGame lottoGame = new LottoGame(new LottoCount());
+		// LottoGame lottoGame = new LottoGame(new LottoCountDto(readMoney()).getLottoCount());
 		//
 		// WinningNumber wn = new WinningNumber(readWinningNumber());
 	}
