+++ src/test/java/lotto/dto/LottoCountDtoTest.java
@@ -1,15 +1,15 @@
-package lotto.domain;
+package lotto.dto;
 
 import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-public class LottoCountTest {
+public class LottoCountDtoTest {
 	@ParameterizedTest
 	@ValueSource(ints = {1500, 4310, 3404, 7146, 10200})
 	void 천원_단위가_아닌_경우(int value) {
-		assertThatThrownBy(() -> new LottoCount(value))
+		assertThatThrownBy(() -> new LottoCountDto(value))
 			.isInstanceOf(IllegalArgumentException.class)
 			.hasMessageContaining("천 원 단위");
 	}
@@ -17,7 +17,7 @@ public class LottoCountTest {
 	@ParameterizedTest
 	@ValueSource(ints = {999, 0, -100})
 	void 한_장도_살_수_없는_경우(int value) {
-		assertThatThrownBy(() -> new LottoCount(value))
+		assertThatThrownBy(() -> new LottoCountDto(value))
 			.isInstanceOf(IllegalArgumentException.class)
 			.hasMessageContaining("부족합니다");
 	}
